Restructure the relationship between InitFromHash and Adapter
InitFromHash is responsible for building objects from QT hashes but the
accessor methods are defined through initializer methods in the Adapter class.
This has created many cases where a class inherites from Adapter just to use
the accessor building methods but many adapter methods were broken because
there is no associated api endpoint so it did not set an api url.

This commit moves the accessor building methods from Adapter into InitFromHash
and removed the unnecessary Adapter Inheritance.
