+++ lib/quick_travel/adjustment.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class Adjustment < Adapter
+  class Adjustment
+    include QuickTravel::InitFromHash
   end
 end
