+++ lib/quick_travel/background_check.rb
@@ -1,7 +1,9 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class BackgroundCheck < Adapter
+  class BackgroundCheck
+    include QuickTravel::InitFromHash
+
     def self.check(options = {})
       get_and_validate('/background_check/check.json', options)
     end
