+++ lib/quick_travel/bed_configuration.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class BedConfiguration < Adapter
+  class BedConfiguration
+    include QuickTravel::InitFromHash
   end
 end
