+++ lib/quick_travel/bed_requirement.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class BedRequirement < Adapter
+  class BedRequirement
+    include QuickTravel::InitFromHash
   end
 end
