+++ lib/quick_travel/client.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class Client < Adapter
+  class Client
+    include QuickTravel::InitFromHash
   end
 end
