+++ lib/quick_travel/contact.rb
@@ -1,4 +1,7 @@
+require 'quick_travel/init_from_hash'
+
 module QuickTravel
-  class Contact < Adapter
+  class Contact
+    include QuickTravel::InitFromHash
   end
 end
