+++ lib/quick_travel/passenger.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class Passenger < Adapter
+  class Passenger
+    include QuickTravel::InitFromHash
   end
 end
