+++ lib/quick_travel/passenger_price_break.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class PassengerPriceBreak < Adapter
+  class PassengerPriceBreak
+    include QuickTravel::InitFromHash
   end
 end
