+++ lib/quick_travel/property_facility.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class PropertyFacility < Adapter
+  class PropertyFacility
+    include QuickTravel::InitFromHash
   end
 end
