+++ lib/quick_travel/room_facility.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class RoomFacility < Adapter
+  class RoomFacility
+    include QuickTravel::InitFromHash
   end
 end
