+++ lib/quick_travel/route_stop.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class RouteStop < Adapter
+  class RouteStop
+    include QuickTravel::InitFromHash
   end
 end
