+++ lib/quick_travel/service.rb
@@ -1,6 +1,7 @@
-require 'quick_travel/adapter'
+require 'quick_travel/init_from_hash'
 
 module QuickTravel
-  class Service < Adapter
+  class Service
+    include QuickTravel::InitFromHash
   end
 end
